+++ packages/core/src/RenderChild.ts
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default interface RenderChild<P> {
-  (props: P): React.ReactElement<P>
-}